+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Rename `NetworkTick` into `RepliconTick` and move it into `server` module.
+
 ## [0.12.0] - 2023-10-01
 
 ### Changed